+++ src/Assets/Code/Runtime/infrastructure/Service/Random/IRandomInterface.cs
@@ -5,7 +5,7 @@ namespace Code.Runtime.infrastructure.Service.Random
 {
     public interface IRandomInterface
     {
-        float Range(float minInclusive, float maxInclusive);
+        int Range(int minInclusive, int maxInclusive);
         T ChooseFromList<T>(List<T> List);
     }
 }
